requirements.txt
langchain
langgraph
langchain-mcp-adapters
langchain-openai
python-dotenv
uv
mcp
mcp[cli]